Commit project foundation
The calculator project is not yet finished, it requires a few tweaks and
some polishing. It already has some basic functionality. Commiting to
have a snapshot of project so far. I plan to add keyboard support,
character limit for the result of equations, rounding of decimals and
handling division by 0.